--- www_src/blocks/all.jsx
@@ -1,5 +1,0 @@
-module.exports = {
-  image: require('./image.jsx'),
-  text: require('./text.jsx'),
-  link: require('./link.jsx')
-};
