+++ www_src/components/el/el.less
@@ -1,0 +1,75 @@
+.el-container {
+  position: absolute;
+  height: 1px;
+  width: 1px;
+  left: 50%;
+  top: 50%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.touch-overlay {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  z-index: 2147483647; /* signed 32bit maxint */
+}
+
+.el {
+  flex-shrink: 0;
+
+  padding: 10px;
+  user-select: none;
+
+  position: relative;
+  z-index: 0;
+
+  &.current {
+    margin: -1px;
+    border: 1px dashed @white;
+
+    &:after {
+      z-index: -1;
+      content: '';
+      display: block;
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      border: 1px dashed @blue;
+    }
+  }
+}
+
+/*********************************************************
+* Types
+*/
+
+.el {
+
+  // Text
+  &.el-text {
+    p {
+      margin: 0;
+    }
+  }
+
+  // Link
+  &.el-link {
+    .btn {
+      margin: 0;
+    }
+  }
+
+  // Image
+  .el-image {
+    img {
+      display: block;
+    }
+  }
+}
+
