+++ www_src/components/el/types/text.jsx
@@ -1,6 +1,6 @@
 var React = require('react');
 var assign = require('react/lib/Object.assign');
-var Spec = require('../lib/spec');
+var Spec = require('../../../lib/spec');
 
 var spec = new Spec('text', assign({
   innerHTML: {
@@ -60,10 +60,6 @@ var Text = React.createClass({
     };
     ['fontFamily', 'color', 'fontWeight', 'fontStyle', 'textDecoration', 'textAlign']
       .forEach(prop => style[prop] = props[prop]);
-
-    if (props.position) {
-      style = assign(style, Spec.propsToPosition(props));
-    }
 
     var content = this.makeEditable(props.innerHTML, style);
     var onPClick = this.activate;
