+++ www_src/components/element-group/element-group.jsx
@@ -1,5 +1,5 @@
 var React = require('react');
-var Positionable = require('../../pages/page/positionable.jsx');
+var El = require('../../components/el/el.jsx');
 var assign = require('react/lib/Object.assign');
 
 var ElementGroup = React.createClass({
@@ -11,15 +11,13 @@ var ElementGroup = React.createClass({
     };
   },
   render: function () {
-    return (<div className="positionables">
+    return (<div className="element-group">
       <div className="deselector" onClick={this.props.onDeselect} />
       {this.props.elements.map((elProps, i) => {
 
         if (!elProps || !elProps.type) return false;
 
         elProps = assign({}, elProps, {
-          parentWidth: this.props.dims.width,
-          parentHeight: this.props.dims.height,
           isCurrent: this.props.currentElement === i,
           interactive: this.props.interactive
         });
@@ -28,7 +26,7 @@ var ElementGroup = React.createClass({
         if (this.props.onTouchEnd) elProps.onTouchEnd = this.props.onTouchEnd(i);
         if (this.props.onUpdate) elProps.onUpdate = this.props.onUpdate(i);
 
-        return <Positionable key={'positionable' + i} {...elProps} />
+        return <El key={'positionable' + i} {...elProps} />
       })}
     </div>);
   }
