+++ www_src/components/element-group/element-group.less
@@ -1,0 +1,9 @@
+/*********************************************************
+* Positionables
+*/
+.element-group {
+  position: relative;
+  overflow: visible;
+  height: 100%;
+  width: 100%;
+}
