+++ www_src/lib/router.jsx
@@ -21,8 +21,8 @@ module.exports = {
     } else {
       r = {
         user: 1,
-        project: 25,
-        page: 94,
+        project: 1,
+        page: 1,
         element: 1
       };
     }
