+++ www_src/lib/touchhandler.js
