+++ www_src/main.less
@@ -16,6 +16,8 @@
 @import "components/alert/alert";
 @import "components/text-input/text-input";
 @import "components/loading/loading";
+@import "components/element-group/element-group";
+@import "components/el/el";
 
 // Pages ----------------------------------------------------------------------
