+++ www_src/pages/element/text-editor.jsx
@@ -2,7 +2,7 @@ var React = require('react/addons');
 var defaults = require('lodash.defaults');
 var render = require('../../lib/render.jsx');
 
-var TextBlock = require('../../blocks/text.jsx');
+var TextBlock = require('../../components/el/types/text.jsx');
 var Range = require('../../components/range/range.jsx');
 var ColorGroup = require('../../components/color-group/color-group.jsx');
 var {CheckboxSet, Radio} = require('../../components/option-panel/option-panel.jsx');
