+++ www_src/pages/page/page.less
@@ -148,97 +148,3 @@
     }
   }
 }
-
-/*********************************************************
-* Positionables
-*/
-.tester {
-  background: white;
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 320px;
-  height: 480px;
-  z-index: 10000;
-  border: 3px solid blue;
-  .tester-inner {
-    position: absolute;
-    height: 1px;
-    width: 1px;
-    left: 50%;
-    top: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    > div {
-      flex-shrink: 0;
-    }
-    &.blue {
-      background: blue;
-    }
-    &.green {
-      background: green;
-    }
-  }
-}
-
-.positionables {
-  position: relative;
-  overflow: visible;
-  height: 100%;
-  width: 100%;
-
-  .positionable-container {
-    position: absolute;
-    height: 1px;
-    width: 1px;
-    left: 50%;
-    top: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-
-  .touch-overlay {
-    position: fixed;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    z-index: 2147483647; /* signed 32bit maxint */
-  }
-
-  .positionable {
-    flex-shrink: 0;
-
-    padding: 10px;
-    user-select: none;
-
-    position: relative;
-    z-index: 0;
-    background: inherit;
-
-    p {
-      margin: 0;
-    }
-
-    &.current {
-      margin: -1px;
-      border: 1px dashed @white;
-
-      &:after {
-        z-index: -1;
-        content: '';
-        display: block;
-        position: absolute;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-        border: 1px dashed @blue;
-      }
-    }
-  }
-
-
-}
